gcp/requirements.txt
numpy
Pillow
pydantic
fastapi
uvicorn
gunicorn
onnx
onnxruntime